+++ grid/Column.php
@@ -1,0 +1,114 @@
+<?php
+/**
+ * @link http://www.euqol.com/
+ * @copyright Copyright (c) 2015 Su Anli
+ * @license http://www.euqol.com/license/
+ */
+
+namespace anli\helper\grid;
+
+use anli\helper\assets\AjaxDeleteAsset;
+use Yii;
+use yii\grid\ActionColumn;
+use yii\helpers\Html;
+use yii\helpers\Url;
+
+/**
+ * This is the base column class.
+ *
+ * @author Su Anli <anli@euqol.com>
+ * @since 2.0.0
+ */
+class Column
+{
+    public $controller = '';
+
+    /**
+     * @var array
+     */
+    public $columns = [];
+
+    /**
+     * @return array
+     */
+    public function all()
+    {
+        return $this->columns;
+    }
+
+    /**
+     * @return string $template
+     * @return mixed
+     */
+    public function actions($template = '{update} {delete}')
+    {
+        $this->columns = array_merge($this->columns, [
+            [
+                'class' => ActionColumn::className(),
+                'controller' => $this->controller,
+                'template' => $template,
+                'buttons' => [
+                    'ajax-update' => function ($url, $model) {
+                        return Html::a(Html::tag('i', '', ['class' => 'glyphicon glyphicon-pencil']),
+                            false,
+                            [
+                                'value' => Url::to(["$this->controller/ajax-update", 'id' => $model->id, 'pjaxId' => 'list-pjax']),
+                                'data-toggle' => 'tooltip',
+                                'title' => 'Update',
+                                'class' => 'showModalButton',
+                                'name' => "update-$this->controller-button",
+                            ]
+                        );
+                    },
+                    'ajax-delete' => function ($url, $model) {
+                        AjaxDeleteAsset::register(Yii::$app->controller->view);
+
+                        return  Html::a('<span  class="glyphicon glyphicon-trash"></span>', false, [
+                            'class' => 'ajaxDelete',
+                            'delete-url' => Url::to(["$this->controller/ajax-delete", 'id' => $model->id]),
+                            'pjax-container' => 'list-pjax',
+                            'name' => "delete-$this->controller-button",
+                        ]);
+                    },
+                    'ajax-copy' => function ($url, $model) {
+                        return Html::a(Html::tag('i', '', ['class' => 'glyphicon glyphicon-copy']),
+                            false,
+                            [
+                                'value' => Url::to(["$this->controller/ajax-copy", 'id' => $model->id, 'pjaxId' => 'list-pjax']),
+                                'data-toggle' => 'tooltip',
+                                'title' => 'Copy',
+                                'class' => 'showModalButton',
+                                'name' => "copy-$this->controller-button",
+                            ]
+                        );
+                    },
+                ],
+                'contentOptions' => ['class' => 'text-right'],
+            ]
+        ]);
+        return $this;
+    }
+
+    /**
+     * @param boolean
+     * @return mixed
+     */
+    public function checkbox($show = true)
+    {
+        if ($show) {
+            $this->columns = array_merge($this->columns, [['class' => '\yii\grid\CheckboxColumn']]);
+            return $this;
+        }
+        return $this;
+    }
+
+    /**
+     * @param mixed
+     * @return mixed
+     */
+    public function raw($value = '')
+    {
+        $this->columns = array_merge($this->columns, $value);
+        return $this;
+    }
+}